requirements.txt
click==8.1.7
joblib==1.3.2
nltk==3.8.1
numpy==1.26.0
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.0.1
pytesseract==0.3.10
regex==2023.10.3
tqdm==4.66.1
